% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAseq.multi.Laplace.R
\name{BMAseq.multi.Laplace}
\alias{BMAseq.multi.Laplace}
\title{BMAseq.multi}
\usage{
BMAseq.multi.Laplace(
  dat.expr.counts,
  dat.pheno,
  var.pool,
  max.nvar,
  interaction = NULL,
  ind.incl.add = NULL,
  cut.BF = 1,
  cut.FDR = 0.05
)
}
\arguments{
\item{dat.expr.counts}{RNA-seq count data matrix (rows for genes and columns for subjects).}

\item{dat.pheno}{Phenotypic data matrix (rows for subjects and columns for variables).}

\item{var.pool}{Variables of interest, a vector.}

\item{max.nvar}{The maximum number of variables in a model.}

\item{interaction}{Specific interaction terms with input format 'A&B', default is Null.}

\item{ind.incl.add}{Indices of additional included model.}

\item{cut.BF}{Bayes factor criterion, default value is 1.}

\item{cut.FDR}{False discovery rate criterion for identifying DE genes, default value is 0.05.}
}
\value{
A list consisting of
\item{dat.expr.logcpm}{Normalized RNA-seq data matrix (rows for genes and columns for subjects).}
\item{weights}{Estimated voom weights.}
\item{dat.pheno.new}{Phenotypic data matrix including new interaction variables, will be same as input dat.pheno if interaction is NULL.}
\item{model.space}{Model space including all possible models.}
\item{post.modelprob}{Posterior model probability for each gene.}
\item{post.incl.modelprob}{Posterior inclusive model probability for each gene associated with main effect of each variables if interaction is NULL; with main, interaction, main or interaction effects if interaction not NULL.}
\item{post.incl.modelprob.JointMain}{Posterior inclusive model probability for each gene associated with joint main effects of variables, will be output if interaction is NULL.}
\item{post.incl.modelprob.Add}{Posterior inclusive model probability for each gene associated with additional inclusive models, will be output if ind.incl.add is not Null.}
\item{index.incl}{Indices of the inclusive models.}
\item{eFDR}{Estimated FDR for each gene associated with main effect of each variables if interaction is NULL; with main, interaction, main or interaction effects if interaction is not NULL.}
\item{eFDR.JointMain}{Estimated FDR for each gene associated with all possible joint effect of variables, will be output if interaction is NULL.}
\item{eFDR.Add}{Estimated FDR for each gene associated with additional inclusive models, will be output if ind.incl.add is not Null.}
\item{summary.nDEG}{A summary table of the number of identified DE gene associated with main effect of each variable if interaction is NULL; with main, interaction, main or interaction effects if interaction is not NULL.}
\item{summary.nDEG.JointMain}{A summary table of the number of identified DE gene associated with joint main effcts of variables, will be output if interaction is NULL.}
\item{summary.nDEG.Add}{A summary table of the number of identified DE gene associated with additional inclusive models, will be output if ind.incl.add is not Null.}
\item{DEG.bestmodel}{DE genes associated with main effect of each variable if interaction is NULL; with main, interaction, main or interaction effects of each variable if interaction is not NULL; and the best model used to identify each DE gene.}
}
\description{
Multivariate analysis using BMAseq approach
}
\details{
Multivariate analysis on RNA-seq count data using BMAseq approach
}
\author{
Lingsong Meng
}
