% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.jointMain.R
\name{get.jointMain}
\alias{get.jointMain}
\title{get.jointMain}
\usage{
get.jointMain(postprob.output, joint = NULL, cut.FDR = 0.05)
}
\arguments{
\item{postprob.output}{The output from BMAseq.multi.postprob function, a list.}

\item{joint}{specific joint main effects with format 'A.B', a vector.}

\item{cut.FDR}{False discovery rate criterion for identifying DE genes, default value is 0.05.}
}
\value{
A list consisting of
\item{post.incl.modelprob.JointMain}{Posterior inclusive model probability for specific joint effect of variables, only use if interaction is NULL.}
\item{eFDR.JointMain}{Estimated FDR for specific joint effect, only use if interaction is NULL.}
\item{summary.nDEG.JointMain}{A summary table of the number of identified DE gene for specific joint effcts of variables, only use if interaction is NULL.}
\item{ind.incl.JointMain}{Indices of the inclusive models for specific joint effcts of variables, only use if interaction is NULL.}
}
\description{
Obtain specific joint main effects output
}
\details{
Obtain specific joint main effects output
}
\author{
Lingsong Meng
}
