% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelspace.R
\name{Modelspace}
\alias{Modelspace}
\title{Modelspace}
\usage{
Modelspace(dat.pheno, var.pool, max.nvar, interaction = NULL)
}
\arguments{
\item{dat.pheno}{Phenotypic data matrix (rows for subjects and columns for variables).}

\item{var.pool}{Variables of interest, a vector.}

\item{max.nvar}{The maximum number of variables in a model.}

\item{interaction}{Specific interaction terms with input format 'A&B', default is Null.}
}
\value{
A list consisting of
\item{model.space}{Model space including all possible models.}
\item{dat.pheno.new}{Phenotypic data matrix including new interaction variables, will be same as input dat.pheno if interaction=NULL.}
\item{var.pool}{Variables of interest, a vector.}
\item{interaction}{Specific interaction terms with format 'A&B', default is Null.}
}
\description{
Create default model space
}
\details{
Create default model space consisting of all possible models
}
\author{
Lingsong Meng
}
